agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "natural-agent"
version = "0.1.0"
description = "Natural language driven automation agent for Ubuntu 22.04+"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Your Name", email = "you@example.com" }]
license = { text = "MIT" }
dependencies = [
  "openai>=1.40.0",
  "typer[all]>=0.12.3",
  "rich>=13.7.1",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.8.2",
  "SQLAlchemy>=2.0.30",
  "pandas>=2.2.2",
  "pyarrow>=16.1.0",
  "beautifulsoup4>=4.12.3"
]

[project.scripts]
agent = "agent.cli:main"

[tool.setuptools]
package-dir = {"agent" = "src/agent"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["agent*"]


